Broadcast .^ for BitArrays
also specialize Bool^Integer while at it
